modified:   include/config.h 	new file:   include/config_checks.h 	modified:   include/config_pins.h 	modified:   include/context_registry.h 	modified:   include/event_bus.h 	modified:   include/manifest.h 	modified:   include/object_classes.h 	modified:   include/ui_draw.h 	new file:   src/config_asserts.cpp 	modified:   src/context_registry.cpp 	modified:   src/event_bus.cpp 	modified:   src/event_router.cpp 	modified:   src/hal_buttons_simple.cpp 	modified:   src/main.cpp

+++ include/context_registry.h
@@ -16,9 +16,9 @@ void registerContext(const char* name, ContextObject* ctx);
 ContextObject* getContextByName(const char* name);
 
 // Retrieves the number of registered contexts
-size_t getRegisteredContextCount();
+uint8_t getRegisteredContextCount();
 
 // Retrieves a context by index (for debug/UI iteration)
-ContextObject* getContextByIndex(size_t index);
+ContextObject* getContextByIndex(uint8_t index);
 
 #endif // CONTEXT_REGISTRY_H
